requirements.txt
torch
transformers
sounddevice
uvicorn[standard]
fastapi
pydantic
numpy
pydantic_settings
python-dotenv
websockets
torchaudio
kokoro
openai
soundfile
hf_xet
pip
faster-whisper
scipy